separate tests for DHTNode and experts-related stuff
